Samlet validering i en mappe
