+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/validation/partials/familierelasjoner.js
@@ -1,9 +1,6 @@
 import * as Yup from 'yup'
 import { requiredDate, requiredString } from '~/utils/YupValidations'
-import {
-	matrikkeladresse,
-	vegadresse,
-} from '~/components/fagsystem/pdlf/form/partials/adresser/validation'
+import { matrikkeladresse, vegadresse } from '~/components/fagsystem/pdlf/form/validation/partials'
 import { testDatoFom, testDatoTom } from '~/components/fagsystem/utils'
 import _get from 'lodash/get'
