+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/validation/partials/index.ts
@@ -1,0 +1,50 @@
+import {
+	bostedsadresse,
+	matrikkeladresse,
+	vegadresse,
+	oppholdsadresse,
+	kontaktadresse,
+	adressebeskyttelse,
+} from './adresser'
+import {
+	doedfoedtBarn,
+	forelderBarnRelasjon,
+	foreldreansvar,
+	sivilstand,
+	nyPerson,
+} from './familierelasjoner'
+import { innflytting } from './innvandring'
+import { kontaktDoedsbo } from './kontaktinformasjonForDoedsbo'
+import { sikkerhetstiltak } from './sikkerhetstiltak'
+import { telefonnummer } from './telefonnummer'
+import { utenlandskId, falskIdentitet } from './identifikasjon'
+import { utflytting } from './utflytting'
+import { fullmakt } from './fullmakt'
+import { statsborgerskap } from './statborgerskap'
+import { vergemaal } from './vergemaal'
+import { tilrettelagtKommunikasjon } from './tilrettelagtKommunikasjon'
+
+export {
+	bostedsadresse,
+	matrikkeladresse,
+	vegadresse,
+	oppholdsadresse,
+	kontaktadresse,
+	adressebeskyttelse,
+	doedfoedtBarn,
+	forelderBarnRelasjon,
+	foreldreansvar,
+	sivilstand,
+	nyPerson,
+	innflytting,
+	kontaktDoedsbo,
+	sikkerhetstiltak,
+	telefonnummer,
+	utenlandskId,
+	falskIdentitet,
+	utflytting,
+	fullmakt,
+	statsborgerskap,
+	vergemaal,
+	tilrettelagtKommunikasjon,
+}
