+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/validation/partials/statborgerskap.js
@@ -1,0 +1,10 @@
+import * as Yup from 'yup'
+import { requiredString } from '~/utils/YupValidations'
+import { testDatoFom, testDatoTom } from '~/components/fagsystem/utils'
+
+export const statsborgerskap = Yup.object({
+	landkode: requiredString.nullable(),
+	gyldigFraOgMed: testDatoFom(Yup.mixed().nullable(), 'gyldigTilOgMed'),
+	gyldigTilOgMed: testDatoTom(Yup.mixed().nullable(), 'gyldigFraOgMed'),
+	bekreftelsesdato: Yup.date().optional().nullable(),
+})
