+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/pdlf/form/validation/partials/utflytting.js
@@ -1,0 +1,8 @@
+import * as Yup from 'yup'
+import { requiredDate, requiredString } from '~/utils/YupValidations'
+
+export const utflytting = Yup.object({
+	tilflyttingsland: requiredString,
+	tilflyttingsstedIUtlandet: Yup.string().optional().nullable(),
+	utflyttingsdato: requiredDate.nullable(),
+})
